refactor(test): group remaining CLI and state coverage
