+++ test/cli/non-interactive-error-classification.test.ts
