+++ test/cli/repro-2666-silent-list-status.test.ts
@@ -28,13 +28,13 @@ import {
   getSandboxInventory,
   type ListSandboxesCommandDeps,
   renderSandboxInventoryText,
-} from "../src/lib/inventory/index.js";
-import { recoverRegistryEntriesWithFallback } from "../src/lib/list-command-deps.js";
-import { resolveGatewayName } from "../src/lib/onboard/gateway-binding.js";
-import { nemoclawStateRoot } from "../src/lib/state/state-root.js";
+} from "../../src/lib/inventory/index.js";
+import { recoverRegistryEntriesWithFallback } from "../../src/lib/list-command-deps.js";
+import { resolveGatewayName } from "../../src/lib/onboard/gateway-binding.js";
+import { nemoclawStateRoot } from "../../src/lib/state/state-root.js";
 import { testTimeoutOptions } from "./helpers/timeouts";
 
-const CLI = path.join(import.meta.dirname, "..", "bin", "nemoclaw.js");
+const CLI = path.join(import.meta.dirname, "../..", "bin", "nemoclaw.js");
 
 function buildDepsWithThrowingRecovery(): ListSandboxesCommandDeps {
   const registryFallback = {
