+++ test/cli/share-command-writable.test.ts
@@ -6,7 +6,7 @@ import os from "os";
 import path from "path";
 import { afterEach, describe, expect, it, vi } from "vitest";
 
-import { checkLocalMountWritable } from "../src/lib/share-command.js";
+import { checkLocalMountWritable } from "../../src/lib/share-command.js";
 
 describe("checkLocalMountWritable (#3192)", () => {
   afterEach(() => {
