+++ test/cli/update.test.ts
