+++ test/state/redact-device-state.test.ts
