+++ test/state/registry.test.ts
@@ -16,7 +16,7 @@ const tmpDir = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-test-"));
 process.env.HOME = tmpDir;
 
 const require = createRequire(import.meta.url);
-const registry = require("../src/lib/state/registry");
+const registry = require("../../src/lib/state/registry");
 
 const regFile = path.join(tmpDir, ".nemoclaw", "sandboxes.json");
 
@@ -600,7 +600,7 @@ describe("registry", () => {
     expect(registry.getDefault()).toBe("beta");
 
     const registryPath = path.resolve(
-      path.join(import.meta.dirname, "..", "src", "lib", "state", "registry.ts"),
+      path.join(import.meta.dirname, "../..", "src", "lib", "state", "registry.ts"),
     );
     const homeDir = path.dirname(path.dirname(regFile));
     const coordinationDir = fs.mkdtempSync(path.join(os.tmpdir(), "registry-restore-race-"));
@@ -1310,7 +1310,7 @@ describe("advisory file locking", () => {
   it("concurrent writers do not corrupt the registry", () => {
     const { spawnSync } = require("child_process");
     const registryPath = path.resolve(
-      path.join(import.meta.dirname, "..", "src", "lib", "state", "registry.ts"),
+      path.join(import.meta.dirname, "../..", "src", "lib", "state", "registry.ts"),
     );
     const homeDir = path.dirname(path.dirname(regFile));
     // Script that spawns 4 workers in parallel, each writing 5 sandboxes
@@ -1411,7 +1411,7 @@ describe("advisory file locking", () => {
       writeMalformedRegistry();
 
       const registryPath = path.resolve(
-        path.join(import.meta.dirname, "..", "src", "lib", "state", "registry.ts"),
+        path.join(import.meta.dirname, "../..", "src", "lib", "state", "registry.ts"),
       );
       const homeDir = path.dirname(path.dirname(regFile));
       const orchestrator = `
