+++ test/state/snapshot.test.ts
@@ -13,8 +13,8 @@ import os from "node:os";
 import path from "node:path";
 import { pathToFileURL } from "node:url";
 import { afterAll, beforeEach, describe, expect, it } from "vitest";
-import { managedStartupE2eProfile } from "../scripts/checks/generate-managed-startup-profile-fixture.mts";
-import { encodeManagedStartupProfile } from "../src/lib/onboard/managed-startup/profile";
+import { managedStartupE2eProfile } from "../../scripts/checks/generate-managed-startup-profile-fixture.mts";
+import { encodeManagedStartupProfile } from "../../src/lib/onboard/managed-startup/profile";
 import { stateDirectoryDiscoverySshSource } from "./helpers/snapshot-state-discovery-fixture";
 
 // Override HOME BEFORE importing sandbox-state — it reads process.env.HOME
@@ -24,7 +24,7 @@ import { stateDirectoryDiscoverySshSource } from "./helpers/snapshot-state-disco
 const ORIGINAL_HOME = process.env.HOME;
 const TMP_HOME = fs.mkdtempSync(path.join(os.tmpdir(), "nemoclaw-snap-naming-"));
 process.env.HOME = TMP_HOME;
-const REPO_ROOT = path.join(import.meta.dirname, "..");
+const REPO_ROOT = path.join(import.meta.dirname, "../..");
 type BackupScalar = string | number | boolean | null | undefined;
 type BackupValue = BackupScalar | BackupManifestOverrides | BackupValue[];
 type SandboxStateModule = typeof import("../src/lib/state/sandbox.js");
